Preserve active OpenTelemetry parent context
