+++ .changeset/preserve-otel-parent-context.md
@@ -1,0 +1,5 @@
+---
+"@effect/opentelemetry": patch
+---
+
+Preserve trace state and locality when adapting active OpenTelemetry parent contexts.
